Move the patient state to the common test package
